refactor: converted WebQA from JavaScript to TypeScript
